requirements.txt
flask==3.0.2
python-dateutil==2.8.2
pandas==2.2.1 